+++ lib/rdf/graph.ex
@@ -145,7 +145,7 @@ defmodule RDF.Graph do
   For a description of the DSL see [this guide](https://rdf-elixir.dev/rdf-ex/description-and-graph-dsl.html).
   """
   defmacro build(opts \\ [], do: block) do
-    Builder.build(block, opts)
+    Builder.build(block, __CALLER__, opts)
   end
 
   @doc """